requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
attrs==25.3.0
cachetools==4.2.2
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
fastapi==0.110.0
frozenlist==1.7.0
geographiclib==2.0
geopy==2.4.1
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
multidict==6.6.3
oauthlib==3.3.1
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dotenv==1.0.1
python-telegram-bot==22.3
pytz==2025.2
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
starlette==0.36.3
tornado==6.5.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
watchdog==6.0.0
yarl==1.20.1
psycopg2-binary==2.9.10
